Summary service tuning for live data

+++ data/store/store.go
@@ -75,5 +75,5 @@ type SummaryRepository interface {
 	GetOutdatedUserIDs(ctx context.Context, page *page.Pagination) ([]string, error)
 	UpdateSummary(ctx context.Context, summary *summary.Summary) (*summary.Summary, error)
 	DistinctSummaryIDs(ctx context.Context) ([]string, error)
-	CreateSummaries(ctx context.Context, summaries []*summary.Summary) (int64, error)
+	CreateSummaries(ctx context.Context, summaries []*summary.Summary) (int, error)
 }
